refactor(cast): move `ZkCast` to dedicated module
